+++ iOS_SDK/OneSignalDevApp/OneSignalDevApp/SwiftTest.swift
@@ -7,15 +7,15 @@
 //
 
 import Foundation
-import OneSignal
-
-class SwiftTest: NSObject {
-    func testSwiftUserModel() {
-        //OneSignal.user()
-        OneSignal.User.onJwtExpired { externalId, completion in
-            completion("test")
-        }
-        let token1 = OneSignal.User.pushSubscription.token
-        let token = OneSignal.User.pushSubscription.token
-    }
-}
+//import OneSignalFramework
+//
+//class SwiftTest: NSObject {
+//    func testSwiftUserModel() {
+//        //OneSignal.user()
+//        OneSignal.User.onJwtExpired { externalId, completion in
+//            completion("test")
+//        }
+//        let token1 = OneSignal.User.pushSubscription.token
+//        let token = OneSignal.User.pushSubscription.token
+//    }
+//}
