+++ iOS_SDK/OneSignalDevApp/OneSignalDevApp/ViewController.h
@@ -29,7 +29,7 @@
 // This project exisits to make testing OneSignal SDK changes.
 
 #import <UIKit/UIKit.h>
-#import <OneSignal/OneSignal.h>
+#import <OneSignalFramework/OneSignalFramework.h>
 
 @interface ViewController : UIViewController <OSInAppMessageDelegate>
