+++ iOS_SDK/OneSignalSDK/Source/OSInAppMessageAction.h
@@ -30,6 +30,7 @@
 #import "OSInAppMessageTag.h"
 #import <OneSignalCore/OneSignalCore.h>
 #import "OSInAppMessagePrompt.h"
+#import "OneSignalInAppMessaging.h"
 
 NS_ASSUME_NONNULL_BEGIN
