+++ iOS_SDK/OneSignalSDK/Source/OSInAppMessageController.m
@@ -43,13 +43,13 @@
         return;
     }
     
-    let request = [OSRequestLoadInAppMessageContent withAppId:OneSignal.appId withMessageId:self.messageId withVariantId:variantId];
+    let request = [OSRequestLoadInAppMessageContent withAppId:[OneSignalConfigManager getAppId] withMessageId:self.messageId withVariantId:variantId];
     
     [OneSignalClient.sharedClient executeRequest:request onSuccess:successBlock onFailure:failureBlock];
 }
 
 - (void)loadPreviewMessageHTMLContentWithUUID:(NSString * _Nonnull)previewUUID success:(OSResultSuccessBlock _Nullable)successBlock failure:(OSFailureBlock _Nullable)failureBlock {
-    let request = [OSRequestLoadInAppMessagePreviewContent withAppId:OneSignal.appId previewUUID:previewUUID];
+    let request = [OSRequestLoadInAppMessagePreviewContent withAppId:[OneSignalConfigManager getAppId] previewUUID:previewUUID];
     
     [OneSignalClient.sharedClient executeRequest:request onSuccess:successBlock onFailure:failureBlock];
 }
