+++ iOS_SDK/OneSignalSDK/Source/OSMigrationController.h
@@ -25,8 +25,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-#import "OneSignal.h"
-
 #ifndef OSMigrationController_h
 #define OSMigrationController_h
